fix(api): Disallow creating courses with non alphanumeric (+-_) and long slug
This fixes a security issue, where a malicious lecturer or admin, or attacker who gained their access
can create a course with a slug that is injected into commands, a worker runs.

Executing just `ffmpeg` in the worker has been considered, but the worker currently relies on the shells
capability to redirect its stdout to a file.

Thanks to @oiidmnk and @lcarilla for reporting the issue!